[fix] Fix problem with types

+++ FakerCore/Generators/ByteGenerator.cs
@@ -4,7 +4,7 @@ public class ByteGenerator : IGenerator
 {
     public object Generate(Type type, GeneratorContext context)
     {
-        return context.Random.Next(byte.MinValue, byte.MaxValue);
+        return (byte)context.Random.Next(byte.MinValue, byte.MaxValue);
     }
 
     public bool CanGenerate(Type type)
